LayoutTests/http/tests/inspector/extensions-resources-redirect.html

<html>
<head>
<script src="inspector-test2.js"></script>
<script src="extensions-resources-test.js"></script>
<script src="extensions-test.js"></script>
<script type="text/javascript">

function extension_doRequest(force, callback, nextTest)
{
    function callbackWrapper(resource)
    {
        var lastCallback = !resource || /\?redirected=true$/.test(resource.request.url);
        if (lastCallback)
            webInspector.resources.onFinished.removeListener(callbackWrapper);
        callback(resource, lastCallback);
    }
    webInspector.resources.onFinished.addListener(callbackWrapper);
    webInspector.inspectedWindow.eval("doRequest(" + force + ")", function(result) {
        if (result)
            callbackWrapper(null);
    });
}

function extension_testGetRedirectResourceContent(nextTest)
{
    function onFinished(resource, lastCallback)
    {
        if (!lastCallback)
            return;
        extension_getResourceByUrl([ /redirect-methods-result.php\?status=302$/ ], function(resource) {
            resource.getContent(onContent)
        });
    }
    function onContent(content, encoding)
    {
        output("content: " + content + ", encoding: " + encoding);
        nextTest();
    }
    extension_doRequest(false, onFinished);
}

function extension_testRedirectResourcesInHAR(nextTest)
{
    function onFinished(resource, lastCallback)
    {
        if (lastCallback)
            webInspector.resources.getHAR(onHAR);
    }
    function onHAR(har)
    {
        var entries = har.entries;
        var urls = [];
        for (var i = 0; i < entries.length; ++i) {
            var url = entries[i].request.url;
            // Workaround for GTK DRT that requests favicon.ico along with the page.
            if (!/\/favicon\.ico$/.test(url))
                urls.push(url);
        }
        urls.sort();
        output("Resources in HAR:\n" + urls.join("\n"));
        nextTest();
    }
    extension_doRequest(false, onFinished);
}

function extension_testRedirectResourceFinished(nextTest)
{
    function onFinished(resource, lastCallback)
    {
        output("Finished resource: " + resource.request.url);
        if (lastCallback)
            nextTest();
    }
    extension_doRequest(true, onFinished);
}

var requestDone = false;

function doRequest(force)
{
    // Only do request once per test suite, to make tests independent on each other.
    // Returns true iff request is alredy done (so the caller shouldn't wait for onFinished).
    if (requestDone && !force)
        return true;
    requestDone = true;
    // We can't use XHR here -- the content for XHRs is pushed from back-end.
    var iframe = document.createElement("iframe");
    iframe.src = "/loading/resources/redirect-methods-result.php?status=302";
    document.body.appendChild(iframe);
    return false;
}

var test = function()
{
    InspectorTest.reloadPage(InspectorTest.runExtensionTests);
}

</script>
</head>
<body onload="runTest()">
<p>Tests WebInspector extension API</p>
</body>
</html>
