LayoutTests/http/tests/inspector/extensions-resources-redirect-expected.txt

Tests WebInspector extension API

Page reloaded.
Started extension.

Running tests...
RUNNING TEST: extension_testGetRedirectResourceContent

content: undefined, encoding: undefined
RUNNING TEST: extension_testRedirectResourceFinished
Finished resource: http://127.0.0.1:8000/loading/resources/redirect-methods-result.php?status=302
Finished resource: http://127.0.0.1:8000/loading/resources/redirect-methods-result.php?redirected=true
RUNNING TEST: extension_testRedirectResourcesInHAR
Resources in HAR:
http://127.0.0.1:8000/inspector/extensions-resources-redirect.html
http://127.0.0.1:8000/inspector/extensions-resources-test.js
http://127.0.0.1:8000/inspector/extensions-test.js
http://127.0.0.1:8000/inspector/inspector-test2.js
http://127.0.0.1:8000/loading/resources/redirect-methods-result.php?redirected=true
http://127.0.0.1:8000/loading/resources/redirect-methods-result.php?redirected=true
http://127.0.0.1:8000/loading/resources/redirect-methods-result.php?status=302
http://127.0.0.1:8000/loading/resources/redirect-methods-result.php?status=302
All tests done.
